refactoring of certain utils + adding selector to get array of events vs map by date as before
