+++ src/client/utils/math.js
@@ -3,4 +3,10 @@ const rnd = (min, max) => {
   return Math.floor(Math.random() * (max - min + 1)) + min
 }
 
-export { rnd }
+const hash = (str) =>  str.split('').reduce((prevHash, currVal) =>
+    ((prevHash << 5) - prevHash) + currVal.charCodeAt(0), 0)
+
+export {
+  rnd,
+  hash
+}
